+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * The `incremental_flattening` attributes are now just named `flattening`,
   although the old name continues to work.
 
+* A simplification rule for branches that return common results (#2526).
+
 ## Removed
 
 ## Changed